Correctly moved event part
