+++ frontend/src/components/admin/eventview/EventTable.vue
@@ -76,9 +76,9 @@
 
 <script setup lang="ts">
 import { ref, onMounted } from 'vue';
-import TableEntry from '../shared/TableEntry.vue';
-import { strapi } from '../../services/strapi';
-import { formatDateTime } from '../DateUtils';
+import TableEntry from "../../shared/TableEntry.vue";
+import { strapi } from '../../../services/strapi';
+import { formatDateTime } from '../../DateUtils';
 
 defineProps<{
   topBarProps?: Record<string, any>
