+++ frontend/src/views/AdminView.vue
@@ -68,7 +68,7 @@ import SidebarItem from '../components/admin/SidebarItem.vue'
 import DualInteractiveBoxes from '../components/admin/DualInteractiveBoxes.vue'
 import EventActionsBox from '../components/EventActionsBox.vue'
 import EventCalendarBox from '../components/EventCalendarBox.vue'
-import EventTable from '../components/admin/EventTable.vue'
+import EventTable from "../components/admin/eventview/EventTable.vue";
 import { h } from 'vue'
 
 const sidebarItems = [
